+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 ## ChangeLog
 
+### [0.2.10] - 2025-07-15
+
+- Added byte slice retrieval methods for node names, attribute names and values, and `PCData`.
+- Some performance optimization for the `Document::check_closing_tag()` method located in the `parser.rs` file.
+- Table display adjustment in README.md (no <style> tag allowed in github).
+
 ### [0.2.9] - 2025-07-14
 
 - Added 75 negative tests for potentially malformed XML content. Some method adjustments in support of malformed content processing.
-- Added CStr retrieval methods for node names, attribute names and values, and PCData when the `use_cstr` feature is enabled.
+- Added `CStr` retrieval methods for node names, attribute names and values, and `PCData` when the `use_cstr` feature is enabled.
 - Tables formatting adjusted in documentation and readme file.
 
 ### [0.2.8] - 2025-07-11